pyproject.toml
[project]
name = "familybond"
version = "0.1.0"
description = "FamilyBond 家话 — 面向异地家庭与老人沟通场景的陪伴型 DeepAgent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.5",
    "langchain>=1.2.3",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.6",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["agent", "storage"]
